Move all functions related to logging into a new component
Keep the old tests calling the function in the old locations in order
to test they are also backwards-compatible. Any new test should go into
the new test_log files.
