+++ .gitignore
@@ -24,6 +24,7 @@ wheels/
 *.egg-info/
 .installed.cfg
 *.egg
+*~
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
